requirements.txt
python>=3.9

customtkinter>=5.2.0
Pillow>=9.5.0
numpy>=1.23.0

torch>=1.13.0
torchvision>=0.14.0
ultralytics>=8.0.0
